1.添加ColumnFamily的查看和删除 2.对columnFamily的创建做线程安全处理 3.添加对应的异常信息 4.其他细节调整

+++ mydb-server/src/main/java/com/mydb/server/model/DeleteModel.java
@@ -2,8 +2,12 @@ package com.mydb.server.model;
 
 import com.mydb.common.beans.CMDMsg;
 import com.mydb.common.beans.Consts;
+import com.mydb.common.beans.DBException;
+import com.mydb.common.beans.Words;
 import com.mydb.server.store.MyStore;
 import static com.mydb.common.beans.DBConfigs.*;
+
+import org.rocksdb.ColumnFamilyHandle;
 
 public class DeleteModel extends BaseModel{
 
@@ -13,6 +17,10 @@ public class DeleteModel extends BaseModel{
 
 	@Override
 	protected Object process() throws Exception {
+		ColumnFamilyHandle cf=getColumnFamily();
+		if(cf==null){
+			throw new DBException(Words.EX_COLUMNFAMILY_NOTEXISTS);
+		}
 		MyStore.db.delete(getColumnFamily(),jbody.getAsString(KEY).getBytes());
 		return Consts.STATUS.OK;
 	}
